test(game-core): move reducer tests down from the web app
The reducer, validators and debug actions live in game-core but were
tested from apps/web/src/state/__tests__ — 37 tests one layer above the
code they exercise. game-core's own coverage floors were set to 28/12/48/27
to accommodate that, leaving the rules engine effectively unguarded when
read on its own.

Move gameReducer, gameReducerSlots and the two DEBUG_* action tests into
packages/game-core/tests and rewrite their imports to the package-relative
style used by the rest of the suite. apps/web/src/state/__tests__ keeps
only the XState machine and session tests, which is what it is for.

game-core: 27 -> 64 tests, coverage 28/12/48/27 -> 91/77/82/91.
Floors raised to 87/73/78/87. Web coverage is unchanged, confirming these
tests never covered apps/web code.
