+++ packages/game-core/vitest.config.ts
@@ -8,15 +8,15 @@ export default defineConfig({
       include: ['src/**/*.ts'],
       // Tests live in tests/, outside src/, so no in-src test globs to exclude.
       exclude: ['src/**/*.d.ts'],
-      // Floors are deliberately low: the reducer/validators are exercised mostly
-      // by the web state tests (apps/web/src/state/__tests__), and that
-      // cross-package coverage is not attributed here. These just ratchet against
-      // erosion of this package's own tests/ suite.
+      // The reducer is now tested here rather than through the web app, so these
+      // floors reflect real coverage of this package by its own tests/ suite.
+      // Set a few points below current coverage to ratchet against erosion
+      // without breaking CI on normal fluctuation. Raise them as coverage improves.
       thresholds: {
-        statements: 28,
-        branches: 12,
-        functions: 48,
-        lines: 27,
+        statements: 87,
+        branches: 73,
+        functions: 78,
+        lines: 87,
       },
     },
   },
